+++ dynamixel_control/src/snake_hardware_interface.cpp
@@ -76,13 +76,13 @@ bool SnakeRobot::startDXL() {
     // Ping the motors
     for (int i = 0; i < num_joints; i++){
         uint16_t model_number = 0;
-        dxl_comm_result = packetHandler->ping(portHandler.get(), IDs[i], &model_number, &dxl_error);
+        dxl_comm_result = packetHandler->ping(portHandler.get(), (uint8_t)IDs[i], &model_number, &dxl_error);
         if (dxl_comm_result != COMM_SUCCESS){
-            ROS_ERROR("Failed to ping Dynamixel ID %d", IDs[i]);
+            ROS_ERROR("Failed to ping Dynamixel ID %d", (uint8_t)IDs[i]);
             ROS_ERROR("Dynamixel error: %s", packetHandler->getRxPacketError(dxl_error));
             return false;
         }else{
-            ROS_INFO("Dynamixel ID %d pinged successfully. Model number: %d", IDs[i], model_number);
+            ROS_INFO("Dynamixel ID %d pinged successfully. Model number: %d", (uint8_t)IDs[i], model_number);
         }
     }
     ROS_INFO_STREAM("All motors are ready");
@@ -129,7 +129,7 @@ bool SnakeRobot::setTorque(int id, bool enable){
 
 bool SnakeRobot::setTorque(bool enable){
     for (int i = 0; i < num_joints; i++){
-        if (!setTorque(IDs[i], enable)){
+        if (!setTorque((uint8_t)IDs[i], enable)){
             return false;
         }
     }
@@ -151,12 +151,12 @@ void SnakeRobot::read(){
         // Read Present Position (length : 4 bytes) and Convert uint32 -> int32
         // When reading 2 byte data (uint8_t)posfrom AX / MX(1.0), use read2ByteTxRx() instead.
         for (int i = 0; i < num_joints; i++){
-            dxl_addparam_result = bulkRead->addParam(IDs[i], control_table.current_position.address, control_table.current_position.length);
+            dxl_addparam_result = bulkRead->addParam((uint8_t)IDs[i], control_table.current_position.address, control_table.current_position.length);
         }
         dxl_comm_result = bulkRead->txRxPacket();
         if (dxl_comm_result == COMM_SUCCESS) {
             for (int i = 0; i < num_joints; i++){
-                int32_t enc = (int32_t)bulkRead->getData(IDs[i], control_table.current_position.address, control_table.current_position.length);
+                int32_t enc = (int32_t)bulkRead->getData((uint8_t)IDs[i], control_table.current_position.address, control_table.current_position.length);
                 ROS_INFO_STREAM("Raw value for Dynamixel motor associated with joint " << i + 1 << " is " << enc);
                 joint_position_[i] = (enc - init_pose[i]) / DYN2RAD;
             }
@@ -167,12 +167,12 @@ void SnakeRobot::read(){
 
         // Read Present Velocity (length : 4 bytes) 
         for (int i = 0; i < num_joints; i++){
-            dxl_addparam_result = bulkRead->addParam(IDs[i], control_table.current_velocity.address, control_table.current_velocity.length);
+            dxl_addparam_result = bulkRead->addParam((uint8_t)IDs[i], control_table.current_velocity.address, control_table.current_velocity.length);
         }
         dxl_comm_result = bulkRead->txRxPacket();
         if (dxl_comm_result == COMM_SUCCESS) {
             for (int i = 0; i < num_joints; i++){
-                joint_velocity_[i] = (int32_t)bulkRead->getData(IDs[i], control_table.current_velocity.address, control_table.current_velocity.length);
+                joint_velocity_[i] = (int32_t)bulkRead->getData((uint8_t)IDs[i], control_table.current_velocity.address, control_table.current_velocity.length);
             }
             bulkRead->clearParam();
         }else{
@@ -180,12 +180,12 @@ void SnakeRobot::read(){
         }
         // Read Present Load (length : 2 bytes)
         for (int i = 0; i < num_joints; i++){
-            dxl_addparam_result = bulkRead->addParam(IDs[i], control_table.current_load.address, control_table.current_load.length);
+            dxl_addparam_result = bulkRead->addParam((uint8_t)IDs[i], control_table.current_load.address, control_table.current_load.length);
         }
         dxl_comm_result = bulkRead->txRxPacket();
         if (dxl_comm_result == COMM_SUCCESS) {
             for (int i = 0; i < num_joints; i++){
-                joint_effort_[i] = (int16_t)bulkRead->getData(IDs[i], control_table.current_load.address, control_table.current_load.length);
+                joint_effort_[i] = (int16_t)bulkRead->getData((uint8_t)IDs[i], control_table.current_load.address, control_table.current_load.length);
             }
             bulkRead->clearParam();
         }else{
@@ -212,7 +212,7 @@ void SnakeRobot::write(ros::Duration dt) {
             param_goal_position[1] = DXL_HIBYTE(DXL_LOWORD(positions[i]));
             param_goal_position[2] = DXL_LOBYTE(DXL_HIWORD(positions[i]));
             param_goal_position[3] = DXL_HIBYTE(DXL_HIWORD(positions[i]));
-            dxl_addparam_result = bulkWrite->addParam(IDs[i], control_table.goal_position.address, control_table.goal_position.length, param_goal_position);
+            dxl_addparam_result = bulkWrite->addParam((uint8_t)IDs[i], control_table.goal_position.address, control_table.goal_position.length, param_goal_position);
             // ROS_INFO_STREAM("Results from bulkWrite->addParam() ->" << dxl_addparam_result);
         }
         dxl_comm_result = bulkWrite->txPacket();
